backend/requirements.txt
# Существующие зависимости
django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
lxml>=4.9
reportlab>=4.0
pymorphy3>=2.0
razdel>=0.4
python-dotenv>=1.0

# AI зависимости для подхода 1 (ruBERT NER)
torch>=2.0.0
transformers>=4.35.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Для подхода 3 (LLM) - опционально
openai>=1.0.0
# Или для локальной LLM:
# llama-cpp-python>=0.2.0